feature: Add CanvasView
